+++ crates/aviutl2-macros/src/into_script_module_return_value.rs
@@ -6,7 +6,7 @@ pub fn into_script_module_return_value(
 
     let fields = match ast.data {
         syn::Data::Struct(syn::DataStruct {
-            fields: syn::Fields::Named(fields),
+            fields: syn::Fields::Named(ref fields),
             ..
         }) => fields,
         _ => {
@@ -29,10 +29,22 @@ pub fn into_script_module_return_value(
             }
         }
     });
+    let Some(first_field) = fields.named.first() else {
+        return Err(syn::Error::new_spanned(
+            ast,
+            "`IntoScriptModuleReturnValue` cannot be derived for structs with no fields",
+        )
+        .to_compile_error());
+    };
+    let first_field_type = &first_field.ty;
 
     let expanded = quote::quote! {
         impl ::aviutl2::module::IntoScriptModuleReturnValue for #ident {
-            fn into_return_values(self) -> ::aviutl2::AnyResult<Vec<::aviutl2::module::ScriptModuleReturnValue>> {
+            type Err = <::std::option::Option<#first_field_type> as ::aviutl2::module::IntoScriptModuleReturnValue>::Err;
+            fn into_return_values(self) -> ::std::result::Result<
+                ::std::vec::Vec<::aviutl2::module::ScriptModuleReturnValue>,
+                Self::Err,
+            > {
                 let mut map = ::std::collections::HashMap::new();
                 #(#push_fields)*
                 ::aviutl2::module::IntoScriptModuleReturnValue::into_return_values(map)