+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- **Breaking:** APIからのResultはanyhow::Resultを使わないように変更
+- **Breaking:** module: IntoScriptModuleReturnValue traitの返り値をanyhow::Errorから変更
+- **Breaking:** module: `&[T]`、`[T]`はVec<T>として返すように変更
 - **Breaking:** `odbg!`、`oprintln!`を削除
 - `log`クレートで本体のログに出力できるように
 - aviutl2-aliasを追加